+++ packages/models-library/src/models_library/function_service/nodes_group.py
